requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
ansi2html==1.9.1
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloud-sql-python-connector==1.6.0
cryptography==42.0.1
dash==2.14.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-mantine-components==0.12.1
dash-table==5.0.0
Flask==3.0.1
frozenlist==1.4.1
google-auth==2.27.0
greenlet==3.0.3
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
multidict==6.0.4
nest-asyncio==1.6.0
numpy==1.26.3
packaging==23.2
pandas==2.2.0
plotly==5.18.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyMySQL==1.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
retrying==1.3.4
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.25
tenacity==8.2.3
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
Werkzeug==3.0.1
yarl==1.9.4
zipp==3.17.0
